+++ apps/consent-ui/src/components/common/Modal/NewModule.scss
@@ -1,0 +1,21 @@
+.overlay {
+	position: fixed;
+	inset: 0;
+	z-index: 1000;
+
+	&::before {
+		content: '';
+		position: absolute;
+		inset: 0;
+		background: theme('colors.grayscale.800');
+		opacity: 0.5;
+	}
+
+	padding: 3rem 1rem 5rem;
+	@media (min-width: theme('screens.sm')) {
+		padding: 5rem 1rem;
+	}
+	@media (min-width: theme('screens.md')) {
+		padding: 5rem 3rem;
+	}
+}
